moved graph into app.js
