+++ src/HomePage/components/Signup.jsx
@@ -12,18 +12,16 @@ const Signup = () => {
         <div className="auth-container">
             <h2>Signup</h2>
             <form className="login-form">
-                <div className="input-container">
-                    <Input
-                        type="email"
-                        placeholder="Email"
-                        value={email}
-                        onChange={(e) => setEmail(e.target.value)}/>
-                    <Input
-                        type="password"
-                        placeholder="Password"
-                        value={password}
-                        onChange={(e) => setPassword(e.target.value)}/>
-                </div>
+                <Input
+                    type="email"
+                    placeholder="Email"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}/>
+                <Input
+                    type="password"
+                    placeholder="Password"
+                    value={password}
+                    onChange={(e) => setPassword(e.target.value)}/>
                 <div className={'button'}>
                     <span className={'text'}>Signup</span>
                 </div>
